refactor(plugin-compiler-alipay): 优化 wxs 转支付宝或 web 的兼容逻辑
